backend/requirements.txt
absl-py==1.3.0
anyio==3.6.2
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
executing==1.1.1
fastapi==0.85.1
fonttools==4.38.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h11==0.14.0
httptools==0.5.0
idna==3.4
ipython==8.5.0
isort==5.10.1
jedi==0.18.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.4
oauthlib==3.2.2
onnx==1.12.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==4.21.8
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sniffio==1.3.0
stack-data==0.5.1
starlette==0.20.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
